fix: add validateIntent, refactor delete added in generic, fix issue with failed received ops
