+++ libs/coin-modules/coin-tezos/src/logic/index.ts
@@ -7,3 +7,4 @@ export { estimateFees } from "./estimateFees";
 export { getBalance } from "./getBalance";
 export { lastBlock } from "./lastBlock";
 export { listOperations } from "./listOperations";
+export { validateIntent } from "./validateIntent";
